@@ -3,18 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
     "@grpc/grpc-js": "^1.6.3",
     "@grpc/proto-loader": "^0.6.9",
     "@reduxjs/toolkit": "^1.8.1",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^14.0.4",
+    "@testing-library/react": "^13.0.1",
+    "@testing-library/user-event": "^14.1.0",
     "@types/jquery": "^3.5.14",
     "async": "~3.2.3",
-    "google-protobuf": "~3.20.0",
+    "google-protobuf": "~3.20.1-rc.1",
     "grpc-web": "~1.3.1",
     "lodash": "~4.17.0",
     "react": "^18.0.0",
+    "react-avatar": "^4.0.0",
+    "react-bootstrap": "^2.2.3",
     "react-dom": "^18.0.0",
     "react-redux": "^7.2.8",
     "react-router-dom": "^6.3.0",
@@ -25,7 +29,7 @@
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3",
     "web-vitals": "^2.1.4",
-    "webpack": "~5.71.0",
+    "webpack": "~5.72.0",
     "webpack-cli": "~4.9.2"
   },
   "scripts": {
