@@ -1,11 +1,10 @@
 import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
 import userReducer from "../features/user/userSlice";
-// import movieReducer from "../features/movie/movieSlice";
 
 export default configureStore({
   reducer: {
     user: userReducer,
-    movie: movieReducer,
+    // movie: movieReducer,
   },
   middleware: getDefaultMiddleware({
     serializableCheck: false,
