@@ -1,112 +1,104 @@
 import styled from "styled-components";
+import { useDispatch, useSelector } from "react-redux";
+import { useHistory } from "react-router-dom";
+import {
+  selectUserName,
+  selectUserPhoto,
+  setUserLoginDetails,
+  setSignOutState,
+} from "../features/user/userSlice";
+import { useEffect } from "react";
+import { FormServiceClient } from "./proto/FormServiceClientPb";
+import $ from "jquery";
+import { LoginRequest, LoginForm } from "./proto/form_pb";
+import "./styles.css";
+import "./normalize.css";
+import Home from './Home'
+import { useNavigate } from "react-router-dom";
 
-const Login = (props) => {
-  return (
-    <Container>
-    <Content>
-      <CTA>
-      <CTALogoOne src="/images/cta-logo-one.svg"></CTALogoOne>
-      <SignUp>GET ALL THERE</SignUp>
-      <Description>Get Premier Access to Raya and the Last Dragon for an additional fee with a Disney+ subscription. As of 03/26/21, the price of Disney+ and The Disney Bundle will increase by $1.</Description>
-      <CTALogoTwo src="/images/cta-logo-two.png"></CTALogoTwo>
-      </CTA>
-      <BgImage />
-    </Content>
-  </Container>
-  )
-}
+const Login = () => {
+  const dispatch = useDispatch();
+  const userName = useSelector(selectUserName);
+  const navigate = useNavigate();
+  const formService = new FormServiceClient('http://localhost:8080', null, null);
 
-const CTALogoTwo = styled.img`
-  margin-bottom:20px;
-  max-width:600px;
-  min-height:1px;
-  display:inline-block;
-  width:100%;
-  vertical-align: bottom;
-`;
-
-const Description = styled.p`
-  color: hsla(0,0%,95.3%,1);
-  font-size:11px;
-  margin: 0 0 24px;
-  line-height:1.5;
-  letter-spacing:1.5px;
-`;
-
-const SignUp = styled.a`
-  font-weight:bold;
-  color: #f9f9f9;
-  background-color: #0063e5;
-  margin-bottom: 12px;
-  width:100%;
-  letter-spacing:1.5px;
-  font-size: 18px;
-  padding:16.5px 0;
-  border: 1px solid transparent;
-  border-radius:4px;
-  &:hover{
-    background-color: #0483ee;
+  class user {
+    constructor(name, email){
+      this.name = name;
+      this.email = email;
+    }
   }
 
-`;
+  // dispatch to redux store
+  const setUser = (user) => {
+    dispatch(
+      setUserLoginDetails({
+        name: user.name,
+        email: user.email,
+      })
+    );
+  };
+  
+  const INTERVAL = 3000;
+  function login(e) {
+      e.preventDefault();
+      
+      var formData = $("#login_form").serializeArray();
+      var email = formData[0].value.trim();
+      var password = formData[1].value.trim();
+      var form = new LoginForm();
+      form.setEmail(email);
+      form.setPassword(password);
+      if (!form)
+        return false;
+      var request = new LoginRequest();
+      request.setForm(form);
+      
+      var call = formService.login(request, {}, function (err, response) {
+        if (err) {
+          console.log(err);
+          return null;
+        } else {
+          var result = document.getElementById("result");
+          if (!response.getSuccess()) {
+              // show the error message in red color
+              result.innerHTML = `<span style="color:red">${response.getMessage()}</span>`;
+          } else {
+              // show success, and redirect to the login page in 10sec
+              const name = "Mingkai"; // TODO: need to be replaced after redefining the protobuf
+              result.innerHTML = `<span style="color:green"><p>${"Successfully logged in!"}</br>${"Redirecting to the home page..."}</p></span>`;
+              setTimeout(() => { setUser(new user(name, email)); navigate('/') }, INTERVAL);
+          }
+        }
+      });
+    }
 
-const Container = styled.section`
-  overflow:hidden;
-  display:flex;
-  flex-direction: column;
-  text-align:center;
-  height: 100vh;
-`;
+  return (
+  <div className="card-wrapper">
+    <div className="card">
+      <h1 className="center">Login</h1>
+      <form name="login_form" id="login_form" >
+        <label htmlFor="email">Email</label>
+        <input type="email" name="email" className="field" placeholder="Enter email" required />
+        
+        <label htmlFor="password">Password</label>
+        <input type="password" name="password" className="field" placeholder="Enter password" required />
+        
+        <div id="result" className="center"></div><br/>
+        <div id="goSignup" className="center">
+         <Link> <a href="/signup" >Create an account</a></Link>
+        </div>
 
-const Content = styled.div`
-  margin-bottom:10vw;
-  width:100%;
-  position:relative;
-  min-height:100vh;
-  box-sizing:border-box;
-  display: flex;
-  justify-content:center;
-  align-items:center;
-  flex-direction:column;
-  padding:80px 40px;
-  height: 100%; 
-`;
-
-const BgImage = styled.div`
-height: 100%;
-background-position:top;
-background-size:cover;
-background-image: url("/images/login-background.jpg");
-background-repeat:no-repeat;
-position:absolute;
-top:0;
-right:0;
-left:0;
-z-index:-1;
-`;
-
-const CTA = styled.div`
-margin-bottom:2px;
-max-width:650px;
-flex-wrap:wrap;
-display:flex;
-flex-direction:column;
-justify-content:center;
-margin-top:0;
-align-items:center;
-text-align:center;
-margin-right:auto;
-margin-left:auto;
-transition-timing-function:ease-out;
-transition:opacity 0.2s;
-width:100%;
-`;
-
-const CTALogoOne = styled.img`
-  margin-bottom:12px;
-  max-width:600px;
-  min-height:1px;
-  display:block;
-  width:100%;
+        <input type="submit" value="Log in" className="btn" onClick={login}/>
+      </form>
+    </div>
+  </div>
+  )
+};
+          
+const Link = styled.div`
+  a {
+    color:blue;
+  }
 `;
 export default Login;
