@@ -3,7 +3,7 @@ import { createSlice } from "@reduxjs/toolkit";
 const initialState = {
   name: "",
   email: "",
-  photo: "",
+  // photo: "",
 };
 
 const userSlice = createSlice({
@@ -11,15 +11,16 @@ const userSlice = createSlice({
   initialState,
   reducers: {
     setUserLoginDetails: (state, action) => {
+      console.log(action);
       state.name = action.payload.name;
       state.email = action.payload.email;
-      state.photo = action.payload.photo;
+      // state.photo = action.payload.photo;
     },
 
     setSignOutState: (state) => {
       state.name = null;
       state.email = null;
-      state.photo = null;
+      // state.photo = null; 
     },
   },
 });
