@@ -3,21 +3,17 @@ import ReactDOM from "react-dom";
 import {createRoot} from "react-dom/client"
 import "./index.css";
 import App from "./App";
-// import reportWebVitals from './reportWebVitals';
-
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>,
-//   document.getElementById('root')
-// );
+import { Provider } from "react-redux";
+import store from "./app/store";
 
 const rootElement = document.getElementById("root");
 const root = createRoot(rootElement);
 
 root.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+      <App />
+    </Provider>
   </React.StrictMode>,
 );
 
